+++ app/src/main/java/com/barryburgle/gameapp/ui/output/section/MonthSection.kt
@@ -2,6 +2,9 @@ package com.barryburgle.gameapp.ui.output.section
 
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.ui.unit.Dp
+import com.barryburgle.gameapp.manager.SessionManager
+import com.barryburgle.gameapp.model.stat.AggregatedPeriod
+import com.barryburgle.gameapp.service.GlobalStatsService
 import com.barryburgle.gameapp.ui.output.OutputCard
 import com.barryburgle.gameapp.ui.output.state.OutputState
 import com.github.mikephil.charting.data.BarEntry
@@ -12,8 +15,21 @@ fun LazyListScope.MonthSection(
     height: Dp,
     width: Dp
 ) {
+    val aggregatedMonthPeriodsList: List<AggregatedPeriod> =
+        SessionManager.createAggregatedPeriodList(
+            state.sessionsByMonth,
+            state.datesByMonth
+        )
+    val aggregatedMonthSessions = SessionManager.getAggregatedSessions(
+        aggregatedMonthPeriodsList,
+        state.lastMonthsShown
+    )
+    val aggregatedMonthDates = SessionManager.getAggregatedDates(
+        aggregatedMonthPeriodsList,
+        state.lastMonthsShown
+    )
     item {
-        state.sessionsByMonth.map { singleMonth ->
+        aggregatedMonthSessions.map { singleMonth ->
             singleMonth.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -32,7 +48,7 @@ fun LazyListScope.MonthSection(
         }
     }
     item {
-        state.sessionsByMonth.map { singleMonth ->
+        aggregatedMonthSessions.map { singleMonth ->
             singleMonth.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -51,7 +67,7 @@ fun LazyListScope.MonthSection(
         }
     }
     item {
-        state.sessionsByMonth.map { singleMonth ->
+        aggregatedMonthSessions.map { singleMonth ->
             singleMonth.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -70,7 +86,7 @@ fun LazyListScope.MonthSection(
         }
     }
     item {
-        state.datesByMonth.map { singleMonth ->
+        aggregatedMonthDates.map { singleMonth ->
             singleMonth.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -89,7 +105,7 @@ fun LazyListScope.MonthSection(
         }
     }
     item {
-        state.sessionsByMonth.map { singleMonth ->
+        aggregatedMonthSessions.map { singleMonth ->
             singleMonth.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -100,7 +116,7 @@ fun LazyListScope.MonthSection(
             OutputCard(
                 height = height,
                 width = width,
-                chartLabel = "Average Index",
+                chartLabel = "Average Session Index",
                 barEntryList = it as List<BarEntry>,
                 integerValues = false,
                 movingAverageWindow = state.movingAverageWindow
@@ -108,7 +124,7 @@ fun LazyListScope.MonthSection(
         }
     }
     item {
-        state.sessionsByMonth.map { weekStat ->
+        aggregatedMonthSessions.map { weekStat ->
             weekStat.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -127,7 +143,7 @@ fun LazyListScope.MonthSection(
         }
     }
     item {
-        state.datesByMonth.map { weekStat ->
+        aggregatedMonthDates.map { weekStat ->
             weekStat.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -146,7 +162,7 @@ fun LazyListScope.MonthSection(
         }
     }
     item {
-        state.sessionsByMonth.map { singleMonth ->
+        aggregatedMonthSessions.map { singleMonth ->
             singleMonth.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -164,8 +180,9 @@ fun LazyListScope.MonthSection(
             )
         }
     }
+    // TODO: consider adding to average ratios precise ratios for the period and std of ratios in the period (do same for weeks)
     item {
-        state.sessionsByMonth.map { singleMonth ->
+        aggregatedMonthSessions.map { singleMonth ->
             singleMonth.periodNumber?.let {
                 BarEntry(
                     it.toFloat(),
@@ -177,6 +194,28 @@ fun LazyListScope.MonthSection(
                 height = height,
                 width = width,
                 chartLabel = "Average Contact Ratio [%]",
+                barEntryList = it as List<BarEntry>,
+                integerValues = true,
+                movingAverageWindow = state.movingAverageWindow
+            )
+        }
+    }
+    item {
+        aggregatedMonthSessions.map { singleMonth ->
+            singleMonth.periodNumber?.let {
+                BarEntry(
+                    it.toFloat(),
+                    GlobalStatsService.computeGenericRatio(
+                        singleMonth.sets.toInt(),
+                        aggregatedMonthDates.get(it).dates.toInt()
+                    ).toFloat()
+                )
+            }
+        }?.let { it ->
+            OutputCard(
+                height = height,
+                width = width,
+                chartLabel = "Date Ratio [%]",
                 barEntryList = it as List<BarEntry>,
                 integerValues = true,
                 movingAverageWindow = state.movingAverageWindow